pyproject.toml
[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
name = "root"
version = "0.0.0"
[tool.poetry.dependencies]
asyncio = "*"
cryptography = "*"
"discord.py" = "*"
emoji = "*"
flask = "*"
json-store-client = "*"
motor = "*"
python = "^3.7"
